# backend/app/openapi.json
{"openapi": "3.0.2", "info": {"title": "new-project-skeleton", "version": "0.1.0"}, "paths": {"/api/v1/users/": {"get": {"tags": ["users"], "summary": "Read Users", "description": "Retrieve users.", "operationId": "read_users_api_v1_users__get", "parameters": [{"required": false, "schema": {"title": "Skip", "type": "integer", "default": 0}, "name": "skip", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer", "default": 100}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Read Users Api V1 Users  Get", "type": "array", "items": {"$ref": "#/components/schemas/User"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["users"], "summary": "Create User", "description": "Create new user.", "operationId": "create_user_api_v1_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/users/{user_id}": {"get": {"tags": ["users"], "summary": "Read User By Id", "description": "Get a specific user by id.", "operationId": "read_user_by_id_api_v1_users__user_id__get", "parameters": [{"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["users"], "summary": "Update User", "description": "Update a user.", "operationId": "update_user_api_v1_users__user_id__put", "parameters": [{"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["users"], "summary": "Delete User By Id", "description": "Update a user.", "operationId": "delete_user_by_id_api_v1_users__user_id__delete", "parameters": [{"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/users/email/{email}": {"get": {"tags": ["users"], "summary": "Read User By Id", "description": "Get a specific user by id.", "operationId": "read_user_by_id_api_v1_users_email__email__get", "parameters": [{"required": true, "schema": {"title": "Email", "type": "string"}, "name": "email", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/utils/test-test": {"get": {"tags": ["utils"], "summary": "Test Test", "operationId": "test_test_api_v1_utils_test_test_get", "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/utils/get-time": {"get": {"tags": ["utils"], "summary": "Available Trays", "description": "Get server time.", "operationId": "available_trays_api_v1_utils_get_time_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Available Trays Api V1 Utils Get Time Get", "type": "string"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "User": {"title": "User", "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}, "is_active": {"title": "Is Active", "type": "boolean", "default": true}, "is_superuser": {"title": "Is Superuser", "type": "boolean", "default": false}, "full_name": {"title": "Full Name", "type": "string"}, "id": {"title": "Id", "type": "integer"}}}, "UserCreate": {"title": "UserCreate", "required": ["email", "password"], "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}, "is_active": {"title": "Is Active", "type": "boolean", "default": true}, "is_superuser": {"title": "Is Superuser", "type": "boolean", "default": false}, "full_name": {"title": "Full Name", "type": "string"}, "password": {"title": "Password", "type": "string"}}}, "UserUpdate": {"title": "UserUpdate", "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}, "is_active": {"title": "Is Active", "type": "boolean", "default": true}, "is_superuser": {"title": "Is Superuser", "type": "boolean", "default": false}, "full_name": {"title": "Full Name", "type": "string"}, "password": {"title": "Password", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}